+++ api/src/services/schema.test.ts
@@ -6,8 +6,8 @@ import knex from 'knex';
 import { createTracker, MockClient, Tracker } from 'knex-mock-client';
 import { afterEach, beforeAll, describe, expect, test, vi } from 'vitest';
 import type { Collection } from '../types/collection.js';
-import { applyDiff } from '../utils/apply-diff.js';
-import { getSnapshot } from '../utils/get-snapshot.js';
+import { applyDiff } from '../utils/schema/apply-diff.js';
+import { getSnapshot } from '../utils/schema/get-snapshot.js';
 import { SchemaService } from './schema.js';
 
 vi.mock('directus/version', () => ({ version: '0.0.0' }));
@@ -16,11 +16,11 @@ vi.mock('../../src/database/index.js', () => {
 	return { __esModule: true, default: vi.fn(), getDatabaseClient: vi.fn().mockReturnValue('postgres') };
 });
 
-vi.mock('../utils/get-snapshot.js', () => ({
+vi.mock('../utils/schema/get-snapshot.js', () => ({
 	getSnapshot: vi.fn(),
 }));
 
-vi.mock('../utils/apply-diff.js', () => ({
+vi.mock('../utils/schema/apply-diff.js', () => ({
 	applyDiff: vi.fn(),
 }));
