+++ api/src/services/schema.ts
@@ -9,12 +9,12 @@ import type {
 } from '@directus/types';
 import type { Knex } from 'knex';
 import getDatabase from '../database/index.js';
-import { applyDiff } from '../utils/apply-diff.js';
-import { getSnapshotDiff } from '../utils/get-snapshot-diff.js';
-import { getSnapshot } from '../utils/get-snapshot.js';
+import { applyDiff } from '../utils/schema/apply-diff.js';
+import { getSnapshotDiff } from '../utils/schema/get-snapshot-diff.js';
+import { getSnapshot } from '../utils/schema/get-snapshot.js';
 import { getVersionedHash } from '../utils/get-versioned-hash.js';
-import { validateApplyDiff } from '../utils/validate-diff.js';
-import { validateSnapshot } from '../utils/validate-snapshot.js';
+import { validateApplyDiff } from '../utils/schema/validate-diff.js';
+import { validateSnapshot } from '../utils/schema/validate-snapshot.js';
 
 export class SchemaService {
 	knex: Knex;
