+++ api/src/utils/schema/apply-diff.test.ts
@@ -12,15 +12,15 @@ import type { Knex } from 'knex';
 import knex from 'knex';
 import { createTracker, MockClient, Tracker } from 'knex-mock-client';
 import { afterEach, beforeEach, describe, expect, it, type MockedFunction, vi } from 'vitest';
-import { snapshotApplyTestSchema } from '../__utils__/schemas.js';
-import { CollectionsService } from '../services/collections.js';
-import { FieldsService } from '../services/fields.js';
-import { RelationsService } from '../services/relations.js';
-import type { Collection } from '../types/collection.js';
+import { snapshotApplyTestSchema } from '../../__utils__/schemas.js';
+import { CollectionsService } from '../../services/collections.js';
+import { FieldsService } from '../../services/fields.js';
+import { RelationsService } from '../../services/relations.js';
+import type { Collection } from '../../types/collection.js';
+import * as getSchema from '../get-schema.js';
 import { applyDiff, isNestedMetaUpdate } from './apply-diff.js';
-import * as getSchema from './get-schema.js';
 
-vi.mock('../cache.js', () => ({
+vi.mock('../../cache.js', () => ({
 	flushCaches: vi.fn(),
 	getCache: vi.fn(() => ({
 		cache: null,
@@ -45,7 +45,7 @@ vi.mock('../cache.js', () => ({
 	})),
 }));
 
-vi.mock('../emitter.js', () => ({
+vi.mock('../../emitter.js', () => ({
 	default: {
 		emitAction: vi.fn(),
 	},
