+++ api/src/utils/schema/apply-diff.ts
@@ -15,17 +15,17 @@ import type { Diff, DiffDeleted, DiffNew } from 'deep-diff';
 import deepDiff from 'deep-diff';
 import type { Knex } from 'knex';
 import { cloneDeep, merge, set } from 'lodash-es';
-import { flushCaches } from '../cache.js';
-import { getHelpers } from '../database/helpers/index.js';
-import getDatabase from '../database/index.js';
-import emitter from '../emitter.js';
-import { useLogger } from '../logger/index.js';
-import { CollectionsService } from '../services/collections.js';
-import { FieldsService } from '../services/fields.js';
-import { RelationsService } from '../services/relations.js';
-import type { Collection } from '../types/index.js';
-import { transaction } from '../utils/transaction.js';
-import { getSchema } from './get-schema.js';
+import { flushCaches } from '../../cache.js';
+import { getHelpers } from '../../database/helpers/index.js';
+import getDatabase from '../../database/index.js';
+import emitter from '../../emitter.js';
+import { useLogger } from '../../logger/index.js';
+import { CollectionsService } from '../../services/collections.js';
+import { FieldsService } from '../../services/fields.js';
+import { RelationsService } from '../../services/relations.js';
+import type { Collection } from '../../types/index.js';
+import { transaction } from '../transaction.js';
+import { getSchema } from '../get-schema.js';
 
 type CollectionDelta = {
 	collection: string;
