+++ api/src/utils/schema/is-changed.ts
@@ -1,0 +1,7 @@
+import type { Diff } from 'deep-diff';
+
+/** Keep only diff entries that actually changed. */
+export function isChanged(entry: { diff: Diff<any>[] | undefined }) {
+	// `deep-diff` returns `undefined` when the two entities are identical.
+	return Array.isArray(entry.diff);
+}
