+++ api/src/utils/schema/is-non-system.ts
@@ -1,0 +1,5 @@
+/** Exclude system collections/fields */
+export function isNonSystem(item: { meta: { system?: boolean | null } | null }) {
+	if (item?.meta?.system === true) return false;
+	return true;
+}
